trained subset embeddings
remade baseline for structured data

created a baseline for embeddings but its very bad. Made a different baseline model but I havent ran it yet because its alte. Will run it and test performance tomorrow.

created a gitignore file since the subset data is rlly big. We can find a way to share it between us somehow thats not git.